Drop support for Go 1.6 and earlier
